SlimeMoldOptim

Documentation for SlimeMoldOptim.

SlimeMoldOptim.slimeMoldAlgo

SlimeMoldOptim.slimeMoldAlgo — Method

     slimeMoldAlgo( nMolds::Int,
                    n_vc::Int,
                    kwargs...;
                    lower::Vector{Float32},
                    upper::Vector{Float32},
                    objFunction::Function,
                    maxiter::Int,
                    ϵ_conv::Float32,
                    z::Float32=0.03f0 )

  minimize objective function objFunction, the solution space is limited by lower and upper bound.
  - The optimization algorithm utilized is the slime mold algorithm.
  - The nMolds parameter governs the number of molds being modeled.
  - n_vc is the number of steps learning parameter vb needs to reach 0.0.
  - maxiter denotes the overall number of algo iteraions, it has to be >= n_vc.

  the number of sparks per firework, in remains constant foreach firework. ϵ_A is the smoothing parameter
  controlling the variance of amplitudes computed foreach fw. C_a ist the upscaling parameter for explosion
  amplitudes. C_r is the downscaling parameter for explosion amplitudes. XPrimary is the feature set of the
  primary algorithm to be tuned. yPrimary is the target set of the primary algorithm. maxiter is the maximum
  number iteraions.  ϵ_conv denotes the convergence parameter.